# backend/api-docs/swagger.yaml
openapi: 3.0.3
info:
  title: FES Stories API
  description: |
    API documentation for FES Stories application.
    
    This API provides endpoints for authentication, health checks, user management, activity tracking, story categories, and story management.
    
    **Base URL:** `http://localhost:8000`
    
    **Authentication:** Currently not required for public endpoints (will be added for protected endpoints)
    
    **Activity Tracking:** All user activities are logged to the database. Users can only view their own activities.
    
    **Story Management:** 
    - Readers can submit stories for review
    - Super admins can review, approve, reject, edit, and delete stories
    - Published stories are available via public endpoint
  version: 1.0.0
  contact:
    name: FES Stories API Support
    email: support@fesstories.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.fesstories.com
    description: Production server (when available)
...
paths:
...
  /api/story-categories/readers:
    get:
      tags:
        - Story Categories
      summary: Get all readers with their category access
      description: |
        Retrieves all readers with their assigned category access. **Super admin only.**
      operationId: getReadersWithAccess
      responses:
        '200':
          description: Successfully retrieved readers with access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadersWithAccessResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/story-categories/readers/{userId}:
    get:
      tags:
        - Story Categories
      summary: Get categories accessible by a reader
      description: |
        Retrieves all categories that a specific reader has access to.
      operationId: getReaderCategories
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Reader user ID
          example: 2
      responses:
        '200':
          description: Successfully retrieved reader categories
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoryCategoryListResponse'
        '403':
          description: User is not a reader
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/story-categories/readers/{userId}/access:
    put:
      tags:
        - Story Categories
      summary: Update reader access to categories
      description: |
        Updates which categories a reader can post stories to. **Super admin only.**
        This replaces all existing access with the provided category IDs.
      operationId: updateReaderAccess
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Reader user ID
          example: 2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReaderAccessRequest'
            example:
              category_ids: [1, 2, 3]
      responses:
        '200':
          description: Reader access updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateReaderAccessResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '403':
          description: User is not a reader
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          description: Error message
          example: "Invalid credentials"
    ValidationError:
      type: object
      required:
        - success
        - message
        - errors
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Validation failed"
        errors:
          type: object
          description: Validation errors grouped by field
          additionalProperties:
            type: array
            items:
              type: string
          example:
            email:
              - "The email must be a valid email address."
            password:
              - "The password field is required."
    StoryCategory:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Category ID
          example: 1
        name:
          type: string
          maxLength: 255
          description: Category name (unique)
          example: "Technology"
        description:
          type: string
          nullable: true
          description: Category description
          example: "Stories about technology and innovation"
        is_active:
          type: boolean
          description: Whether the category is active
          example: true
        created_at:
          type: string
          format: date-time
          description: ISO 8601 timestamp when category was created
          example: "2025-12-02T10:30:00+00:00"
        updated_at:
          type: string
          format: date-time
          description: ISO 8601 timestamp when category was last updated
          example: "2025-12-02T10:30:00+00:00"
    StoryCategoryListResponse:
      type: object
      required:
        - success
        - categories
      properties:
        success:
          type: boolean
          example: true
        categories:
          type: array
          items:
            $ref: '#/components/schemas/StoryCategory'
...
    ReaderWithCategories:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Reader user ID
          example: 2
        name:
          type: string
          description: Reader name
          example: "John Doe"
        email:
          type: string
          format: email
          description: Reader email
          example: "john@example.com"
        categories:
          type: array
          description: Categories accessible by this reader
          items:
            type: object
            properties:
              id:
                type: integer
                format: int64
                example: 1
              name:
                type: string
                example: "Technology"
              description:
                type: string
                nullable: true
                example: "Stories about technology"
    ReadersWithAccessResponse:
      type: object
      required:
        - success
        - readers
      properties:
        success:
          type: boolean
          example: true
        readers:
          type: array
          items:
            $ref: '#/components/schemas/ReaderWithCategories'
    UpdateReaderAccessRequest:
      type: object
      required:
        - category_ids
      properties:
        category_ids:
          type: array
          description: Array of category IDs the reader should have access to
          items:
            type: integer
            format: int64
          example: [1, 2, 3]
    UpdateReaderAccessResponse:
      type: object
      required:
        - success
        - message
        - categories
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Reader access updated successfully"
        categories:
          type: array
          items:
            $ref: '#/components/schemas/StoryCategory'
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication (to be implemented)